refactor: restructure core node modules
