+++ core/node.go
@@ -1,19 +1,137 @@
 package core
 
-import "synnergy/nodes"
+import (
+	"errors"
+	"fmt"
+)
 
-// NodeAdapter adapts the consensus Node to the generic nodes.NodeInterface.
-type NodeAdapter struct {
-	node *Node
-	*BaseNode
+// Node represents a participant in the network.
+type Node struct {
+	ID            string
+	Addr          string
+	Ledger        *Ledger
+	Consensus     *SynnergyConsensus
+	VM            *SNVM
+	Mempool       []*Transaction
+	Blockchain    []*Block
+	Stakes        map[string]uint64
+	Slashed       map[string]bool
+	MaxTxPerBlock int
 }
 
-// NewNodeAdapter wraps the provided Node with a BaseNode implementing nodes.NodeInterface.
-func NewNodeAdapter(n *Node) *NodeAdapter {
-	return &NodeAdapter{
-		node:     n,
-		BaseNode: NewBaseNode(nodes.Address(n.ID)),
+// NewNode creates a new node instance.
+func NewNode(id, addr string, ledger *Ledger) *Node {
+	return &Node{
+		ID:         id,
+		Addr:       addr,
+		Ledger:     ledger,
+		Consensus:  NewSynnergyConsensus(),
+		VM:         NewSNVM(),
+		Mempool:    []*Transaction{},
+		Blockchain: []*Block{},
+		Stakes:     make(map[string]uint64),
+		Slashed:    make(map[string]bool),
 	}
 }
 
-var _ nodes.NodeInterface = (*NodeAdapter)(nil)
+// AddTransaction validates and adds a transaction to the mempool.
+func (n *Node) AddTransaction(tx *Transaction) error {
+	if err := n.ValidateTransaction(tx); err != nil {
+		return err
+	}
+	n.Mempool = append(n.Mempool, tx)
+	return nil
+}
+
+// ValidateTransaction checks if a transaction is well-formed and the sender has
+// sufficient balance.
+func (n *Node) ValidateTransaction(tx *Transaction) error {
+	// Ensure the fee is considered with the amount using explicit uint64
+	// arithmetic. This guards against future changes to transaction field
+	// types that might otherwise introduce float arithmetic.
+	if n.Ledger.GetBalance(tx.From) < uint64(tx.Amount+tx.Fee) {
+		return errors.New("insufficient funds")
+	}
+	return nil
+}
+
+// MineBlock packages the current mempool into a sub-block and mines a block.
+func (n *Node) MineBlock() *Block {
+	if len(n.Mempool) == 0 {
+		return nil
+	}
+	validator := n.Consensus.SelectValidator(n.eligibleStakes())
+	if validator == "" {
+		return nil
+	}
+	sb := NewSubBlock(n.Mempool, validator)
+	if !sb.VerifySignature() {
+		return nil
+	}
+	n.Mempool = nil
+	prevHash := ""
+	if len(n.Blockchain) > 0 {
+		prevHash = n.Blockchain[len(n.Blockchain)-1].Hash
+	}
+	block := NewBlock([]*SubBlock{sb}, prevHash)
+	n.Consensus.MineBlock(block, 3)
+	var totalFees uint64
+	for _, tx := range sb.Transactions {
+		totalFees += tx.Fee
+		_ = n.Ledger.ApplyTransaction(tx)
+	}
+	n.Blockchain = append(n.Blockchain, block)
+
+	dist := DistributeFees(totalFees)
+	pool := AdjustForBlockUtilization(dist.ValidatorsMiners, len(sb.Transactions), n.MaxTxPerBlock)
+	weights := map[string]uint64{validator: n.Stakes[validator]}
+	weights[n.ID] = 1
+	shares := ShareProportional(pool, weights)
+	contract := NewFeeDistributionContract(n.Ledger)
+	contract.Distribute(shares)
+	return block
+}
+
+const MinStake uint64 = 1
+
+// SetStake assigns stake to an address for validator selection while enforcing a minimum.
+func (n *Node) SetStake(addr string, amount uint64) error {
+	if amount < MinStake {
+		return fmt.Errorf("stake below minimum: %d", amount)
+	}
+	n.Stakes[addr] = amount
+	return nil
+}
+
+func (n *Node) eligibleStakes() map[string]uint64 {
+	eligible := make(map[string]uint64)
+	for addr, s := range n.Stakes {
+		if s >= MinStake && !n.Slashed[addr] {
+			eligible[addr] = s
+		}
+	}
+	return eligible
+}
+
+// ReportDoubleSign slashes a validator for double signing.
+func (n *Node) ReportDoubleSign(addr string) {
+	n.slash(addr)
+}
+
+// ReportDowntime slashes a validator for downtime.
+func (n *Node) ReportDowntime(addr string) {
+	n.slash(addr)
+}
+
+// Rehabilitate removes slashed status from a validator.
+func (n *Node) Rehabilitate(addr string) {
+	delete(n.Slashed, addr)
+}
+
+func (n *Node) slash(addr string) {
+	if stake, ok := n.Stakes[addr]; ok {
+		penalty := stake / 2
+		n.Stakes[addr] = stake - penalty
+		n.Slashed[addr] = true
+	}
+}
